Create A 126 : Precedence
So far we've worked with individual operators, but how do they behave when used together?

Every operator has its own priority or precedence. An operator with higher precedence will be applied before an operator with a lower priority. Precedence is the BODMAS of computing in C. Associativity indicates given two operators of same precedence, in what manner will they be applied i.e. from left to right or from right to left.

Here are the precedence and associativity tables : http://www.difranco.net/compsci/C_Operator_Precedence_Table.htm

Below is the code to take as input four numbers : A, B, C and D and print the value of (A + B) * ( C+ D), however it was written by someone who did not understand operator precedence so it gives the wrong answer. Your task is to fix the code so that it prints the correct answer.

Sample Input 0

1 2 3 4
Sample Output 0

21