Create A 117 : The Address Of Operator
The address of operator or the ampersand '&' in C is used to refer to the address of the memory location where the variable is stored. For example :

int var = 5;
&var will now denote the address of the memory where variable var with its value 5 is stored. Every standard variable has its own memory.

This is precisely why we use ampersand while taking input using scanf. We give scanf the exact location in memory to store the input values to. Scanf("%d", &var) will give the address of var, and hence an integer inputted by the user will be stored at that address.

Fix the code below to make it input one number and print it to the screen.

Sample Input 0

55
Sample Output 0

55