Create A 115 : The Double Data-Type
A double is to float as a long long int is to int. It is a floating data type that occupies twice as much memory as a float and is much more precise. It can hold larger values.

A double can be declared using the keyword double.

Token '%lf' is used in scanf and printf statements for double data types.

To print x digits after the decimal for a double use '%.xlf'.

For this task you must write a program to input two variables to type double and print their division upto 12 decimal places.