Create A 113 : Modifiers
An integer can store values ranging from around -2x10^9 to 2x10^9. But what if we want to store bigger values?

C supports modifiers using which we can vary the size of the datatypes that we want.

For integers, the most common modifiers are :
- unsigned : supports positive integers only
- short : smaller range than actual integer
- long long : larger range than actual integer

Your task is to write a program that reads two long long integers and print its sum.

Note that '%lld' is used while scanning and printing long long integers.

NOTE : You are expected to fix any other errors that you come across.