Create A 109 : User Input
So far we are assigning values to our variables by ourselves. An alternative to this is having the user input the values for our variables.

We can scanf the user inputs using the function scanf which works in similar fashion as printf. The syntax to input an integer variable using scanf is :
scanf("%d", &);

Example for one variable var1 : scanf("%d", &var1);

Example for two variables : scanf("%d%d", &var1, &var2);

For this task, you are to write a code to scan one integer using scanf and print it using printf.

NOTE : You are expected to fix any other errors that you come across.

Sample Input 0

32
Sample Output 0

32