Create A 104 : Declare a Variable
In math, specifically algebra you denoted variables by characters and they could hold any value.

Same is the case with C but the only change is, you can name a variable using as many alphabets/digits as you want as long as the name starts with a letter or an underscore.

You need to inform the C compiler what type of value your variable will be holding. For now, lets try to store a number. In C, a number can be stored in a variable of type integer.

An integer can be declared as follows : int ; and this type of statement is called a declaration statement. For example : int var1; int var2; etc.

Another type of statement is an executable statement, which is a statement that carries out an operation. For example : = ; (like var1 = 5; or var2 = 15;)

For this task you are to write a C code to use a declaration statement to declare a variable called num and to use an executable statement to assign the value '5' to it.

NOTE : You are expected to fix any other errors that you come across.

Sample Output 0

5