Create A 102 : Comments in the Code
You might've noticed certain lines after a double slash '//' or after a '/*'. In C language these lines are called comments. They are not a part of code but they are used to relay information about the code.

'//' is used for single line comments.

'/' and '/' is used for multi-line comments.

The following program has been commented incorrectly. Correct the commenting so it prints “Hello World!”.

NOTE : You are expected to fix any other errors that you come across.

Sample Output 0

Hello World!