Create B 109 : The Break Statement
Sometimes you may want to break the loop before it has completed the number of iterations desired in the beginning because the information that you needed has already been gained. the break keyword does that for you.

Usage: break;

The break keyword will break the loop whenever executed. No more statements after the break statement will be executed and the execution resumes at the line after the loop ends.

The code below tries to find a prime number greater than N. But the loop does not have a condition so it never ends. Break it using break; once it has found a prime number.

Sample Input 0

61
Sample Output 0

67