Create Task 4.5 : Relationship Between Array and Pointers
We learned that an array is a contiguous block of memory and that its name corresponds to the address of the start of the block. A pointer can be allocated memory to function as an array. So are arrays and pointers same? Functionally, yes, they are same.

They only have one essential difference and that is that an array will always point towards a fixed location while a pointer being a variable that stores locations can point towards different locations i.e. the address that a pointer points to can be changed while the same cannot be done for an array.

For this task you are to implement a function char* strrev(char* str) that accepts a string and reverses it, returning the pointer to the reversed string.

Hint : The strlen function may come in handy here.

Sample Input 0

Anarasumanara
Sample Output 0

aranamusaranA